Add BZNet methods to modify input Q and data
If the provided Q points are outside of the first Brillouin zone they are not directly usable.
Instead they must be translated into the first Brillouin zone and, possibly, rotated into the irreducible wedge.
This translation probably effects the passed data, but this can not be accounted for without asking for a callback function that can calculate the change.

Even if all passed Q are inside of an irreducible part of the first Brillouin zone, they may not lie in the same irBz that brille finds.
So rotating to find an equivalent irreducible Q inside of the irBz polyhedron is necessary.
This rotation must be done to both Q and the passed data.
New BrillouinZoneNet3 constructors perform the necessary rotation and re-intialise the held Net object accordingly.